Speedup execution and normalize configuration (#77)
This is a rewrite of how we handle environment variables and subshells. This means that old configuration will no longer work. To fix this backup and remove ${HOME}/.config/sbp and change the casing of 'sbp_path' to 'SBP_PATH' in your ${HOME}/.bashrc. These changes were nessesary to make the code readable when moving away from using subshells.